requirements.txt
flask
requests
psutil
python-dotenv
google-generativeai
praw
openai
